v2.4

+++ search-server/request_queue.h
@@ -29,6 +29,6 @@ private:
     const SearchServer& search_server_;
     const static int min_in_day_ = 1440;
     
-    unsigned long long current_time_;
+    uint64_t current_time_;
     int no_results_requests_;
 };
